refactor(webui): separate desktop controller and view
